Rename not-private symbols
Rename not-private symbols by replacing them with some variation of
their fully qualified name. The renaming is done in two steps.

In RefChecks, the notPrivate flag is set. In Literalize symbols with this flag are renamed.
Literalize was picked because it was not yet an sym transformer. ALternatively, we could
have a separate phas ethat transforms only symbols, not trees.